AUS-4097: Moved filters from LH sidebar to active layers
