+++ src/app/menupanel/activelayers/activelayerspanel.component.html
@@ -1,6 +1,7 @@
 <div class="select-layers-message" *ngIf="getActiveLayers().length===0">
-    Select layers below to add to map
+    Click on "Browse" on map to add layers
 </div>
+
 <div class="activeLayersBoundary" cdkDropList (cdkDropListDropped)="layerDropped($event)">
     <li class="activeLayerItem" *ngFor="let layer of getActiveLayers()" cdkDrag cdkDragLockAxis="y" cdkDragBoundary=".activeLayersBoundary" [cdkDragDisabled]="getActiveLayers().length <= 1">
         <div *ngIf="getActiveLayers().length > 1" cdkDragHandle>
@@ -8,6 +9,7 @@
         </div>
         <div class="activeLayerContent">
             <div>
+                <!-- Layer name, status and warning icons -->
                 <u>{{ layer.name }}
                     <span *ngIf="getUILayerModel(layer.id).statusMap.getRenderStarted()" class="float-right project-percentage hasEvent light-blue">
                         <i *ngIf="!getUILayerModel(layer.id).statusMap.getRenderComplete()" class="fa fa-spin fa-spinner"></i>
@@ -18,15 +20,19 @@
             </div>
             <div style="clear: right" *ngIf="(showOpacitySlider(layer) && getUILayerModel(layer.id).statusMap.getRenderStarted()) || hasLegend(layer)" class="opacity-slider-panel d-flex" (click)="$event.stopPropagation()">
                 <div class="opacity-label">Opacity {{ getUILayerModel(layer.id).opacity }}%&nbsp;</div>
+                <!-- Opacity slider -->
                 <mat-slider class="opacity-slider" min="0" max="100"  (input)="layerOpacityChangeValue($event, layer)">
                     <input matSliderThumb
                         [(ngModel)]="getUILayerModel(layer.id).opacity">
                 </mat-slider>
                 <div style="display: flex;" *ngIf="hasLegend(layer)">
+                    <!-- Display legend button -->
                     <button class="btn btn-primary btn-sm legend"  [disabled]="isLegendShown(layer.id)" (click)="showLegend(layer)">Legend</button>
+                    <!-- Display status button -->
                     <button class="btn btn-primary btn-sm status"  (click)="openStatusReport(getUILayerModel(layer.id)); $event.stopPropagation();">Status</button>
                 </div>
             </div>
+            <!-- Map split controls -->
             <div class="split-panel" *ngIf="getShowSplitMapButtons(layer)" (click)="$event.stopPropagation()">
                 Split Direction&nbsp;
                 <div *ngIf="getApplicableSplitLayer(layer)" class="split-button-panel">
@@ -38,6 +44,48 @@
                     <button class="btn btn-split btn-primary" [disabled]="true">Layer not supported</button>
                 </div>
             </div>
+
+            <div class="sidebar-card-menu-show">
+                <div class="card card-info card-with-tabs layer-card animated slideInRight">
+                    <div class = "rh_info_wrap">
+                        <div class="card-header">
+                            <ul id="card-tab" class="nav nav-tabs">
+                                <!-- Layer Styling tab -->
+                                <li (click)="selectTabPanel(layer.id,'filterpanel')" [ngClass]="{'active': getUILayerModel(layer.id) && getUILayerModel(layer.id).tabpanel.filterpanel.expanded}">
+                                    <a data-toggle="tab">
+                                        <span class="d-none d-sm-inline">Layer Styling</span>
+                                    </a>
+                                </li>
+                                <!-- Layer Download tab -->
+                                <li *ngIf="isMapSupportedLayer(layer) && isDownloadSupportedLayer(layer)" (click)="selectTabPanel(layer.id,'downloadpanel')" [ngClass]="{'active': getUILayerModel(layer.id) && getUILayerModel(layer.id).tabpanel.downloadpanel.expanded}" class="d-none d-md-block">
+                                    <a data-toggle="tab">
+                                        <span>Download</span>
+                                    </a>
+                                </li>
+                            </ul>
+                        </div>
+                        <div class="top-layer-btn-panel">
+                            <!-- Add bookmark button-->
+                            <button *ngIf="isUserLoggedIn() && !isLayerBookmarked(layer.id)" (click)="addLayerBookmark(layer.id)" type="button" class="btn btn-sm btn-info top-layer-btn" title="Bookmark Layer"><i class="fa fa-sm fa-bookmark-o"></i></button>
+                            <!-- Remove bookmark button -->
+                            <button *ngIf="isUserLoggedIn() && isLayerBookmarked(layer.id)" (click)="removeLayerBookmark(layer.id)" type="button" class="btn btn-sm btn-info top-layer-btn" title="Remove Layer Bookmark"><i class="fa fa-sm fa-bookmark"></i></button>
+                            <!-- Info button -->
+                            <button (click)="displayRecordInformation(layer)" type="button" class="btn btn-sm btn-info top-layer-btn" title="Information" [ngClass]="{'active': getUILayerModel(layer.id) && getUILayerModel(layer.id).tabpanel.infopanel.expanded}">Info</button>
+                        </div>
+                    </div>
+                    <div id="card-tab-content" class="tab-content">
+                        <!-- Panel for layer styling -->
+                        <div class="tab-pane fade" [ngClass]="{'show active': getUILayerModel(layer.id) && getUILayerModel(layer.id).tabpanel.filterpanel.expanded}">
+                            <app-filter-panel *ngIf="getUILayerModel(layer.id) && getUILayerModel(layer.id).tabpanel.filterpanel.expanded" [layer]=layer></app-filter-panel>
+                        </div>
+                        <!-- Panel for layer download -->
+                        <div class="tab-pane fade" [ngClass]="{'show active': getUILayerModel(layer.id) && getUILayerModel(layer.id).tabpanel.downloadpanel.expanded}">
+                            <app-download-panel *ngIf="getUILayerModel(layer.id) && getUILayerModel(layer.id).tabpanel.downloadpanel.expanded" [layer]=layer></app-download-panel>
+                        </div>
+                    </div>
+                </div>
+            </div>
+            
         </div>
     </li>
 </div>
