+++ src/app/menupanel/custompanel/layergroup.component.html
@@ -34,7 +34,11 @@
 					<div class = "rh_info_wrap">
 						<div class="card-header">
 							<ul id="card-tab" class="nav nav-tabs float-right">
-								<li (click)="selectTabPanel(layer.id,'filterpanel')" [ngClass]="{'active': getUILayerModel(layer.id).tabpanel.filterpanel.expanded}"><a data-toggle="tab"><span class="d-none d-sm-inline">Map Layer</span></a></li>
+								<li (click)="selectTabPanel(layer.id,'filterpanel')" [ngClass]="{'active': getUILayerModel(layer.id).tabpanel.filterpanel.expanded}">
+									<a data-toggle="tab">
+										<span class="d-none d-sm-inline">Map Layer</span>
+									</a>
+								</li>
 							</ul>
 						</div>
 						<div class="rh_info">
