--- src/app/menupanel/layerpanel/layerpanel.component.scss
@@ -1,68 +1,0 @@
-@import '../../globals';
-
-:host {
-
-    .filter-text{
-        width: 100%;
-    }
-    .filter-btn {
-        margin-left: auto;
-    }
-
-    ::ng-deep {
-        fieldset.show-fieldset-borders {
-            margin: 0 0 1em 0;
-            padding: 0 1.4em 0.5em 1.4em;
-        }
-
-        fieldset legend {
-            width: 50%;
-        }
-
-        .tab-content {
-            padding: 15px 15px 6px 15px;
-        }
-
-        .mat-ripple {
-            display: none !important;
-        }
-
-        .mdc-slider {
-            height: 24px;
-            margin: 0 24px;
-            min-width: 92px;
-        }
-
-        .mdc-slider .mdc-slider__thumb {
-            left: -12px;
-            height: 24px;
-            width: 24px;
-        }
-    }
-
-    .top-layer-btn-panel {
-        display:flex;
-        position:relative;
-        margin-left:auto;
-        right:16px;
-
-        .top-layer-btn {
-            position: relative;
-            top: 16px;
-            height: fit-content;
-            margin-left: 5px;
-        }
-    }
-
-    .show-bookmark-btn {
-        position: relative;
-        top: 5px;
-    }
-
-    .bookmark-icon {
-        color: $auscope-teal;
-        margin-right: 4px;
-        vertical-align: middle;
-    }
-
-}
