+++ src/app/portal/portal.component.html
@@ -57,9 +57,6 @@
         <ul class="card-sub" appDataExplorer></ul>
     </h5>
     <!-- END Data Explorer -->
-    
-    <hr/>
-    <li appPortalDetailsPanel></li>
 </nav>
 <!-- END sidebar -->
