+++ src/app/services/ui/layer-manager.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { CsMapService, LayerModel, ManageStateService, ResourceType, UtilitiesService } from '@auscope/portal-core-ui';
+import { CsMapService, LayerModel, ManageStateService } from '@auscope/portal-core-ui';
 import { AdvancedComponentService } from './advanced-component.service';
 import { LegendUiService } from '../legend/legend-ui.service';
 import { UILayerModelService } from './uilayer-model.service';
@@ -15,16 +15,23 @@ declare let gtag: Function;
 @Injectable()
 export class LayerManagerService {
 
-  constructor(private csMapService: CsMapService, private manageStateService: ManageStateService, private uiLayerModelService: UILayerModelService,
-              private advancedComponentService: AdvancedComponentService, private legendUiService: LegendUiService) {}
+  constructor(private csMapService: CsMapService, private manageStateService: ManageStateService,
+              private uiLayerModelService: UILayerModelService,
+              private advancedComponentService: AdvancedComponentService,
+              private legendUiService: LegendUiService) {}
 
   /**
-   * Add a layer
+   * Add a layer - this is the generic function to add a layer to the map
+   *  - adds the layer to the Cesium map
+   *  - updates the map state service
+   *  - sets up the optional and mandatory filters
+   * 
    *
-   * @param layer
-   * @param optionalFilters
-   * @param layerFilterCollection
-   * @param layerTime
+   * @param layer LayerModel object
+   * @param optionalFilters layer's optional filters that have been selected already
+   * @param layerFilterCollection the layer's filter collection, 
+   *         i.e. mandatory filters and optional filters that can be selected
+   * @param layerTime time range to display
    *
    * TODO: FilterPanel is only place bounding box filter can currently be set, better to shift flag to FilterService
    * and apply here when it's needed to adding from SearchPanel etc. will apply filter as well
@@ -53,7 +60,7 @@ export class LayerManagerService {
       Object.assign(param, advancedFilter.getCallParams());
     }
 
-    // Remove filters without values
+    // Remove filters without values from parameter list
     param.optionalFilters = param.optionalFilters.filter(f => this.filterHasValue(f));
     for (const optFilter of param.optionalFilters) {
       if (optFilter['options']) {
@@ -64,10 +71,25 @@ export class LayerManagerService {
     // Remove any existing legends in case map re-added with new style
     this.legendUiService.removeLegend(layer.id);
 
-    // Add layer to map in Cesium
+    // Transfer mandatory filters from the 'layerFilterCollection' input to the 'layer' object
+    if (layer?.filterCollection?.mandatoryFilters && 
+        layerFilterCollection?.mandatoryFilters) {
+      for (const layerFilt of layer.filterCollection.mandatoryFilters) {
+        for (const mandFilt of layerFilterCollection.mandatoryFilters) {
+          if (layerFilt.label === mandFilt.label) {
+            // Set value of filter
+            layerFilt.value = mandFilt.value;
+          }
+        }
+      }
+    }
 
+    // Add layer to map in Cesium
     this.csMapService.addLayer(layer, param);
-    
+
+    // Update the optional filter display
+    this.csMapService.updateFilterDisplay(layer.id, optionalFilters);
+
     // Add a new layer in the layer state service
     this.manageStateService.addLayer(
       layer.id,
