+++ src/app/toppanel/help-menu/help-menu.component.html
@@ -3,4 +3,6 @@
     <h5 mat-menu-item><a href="http://www.auscope.org.au/" target="href:blank"><em class="ti-home menu-icon"></em><span>&nbsp;About</span></a></h5>
     <h5 mat-menu-item><a href="https://github.com/AuScope/AuScope-Portal-UI/raw/master/docs/AuScopeUserGuide.pdf" target="href:blank"><em class="ti-info-alt menu-icon"></em><span>&nbsp;User Guide</span></a></h5>
     <h5 mat-menu-item><a href="mailto:cg-admin@csiro.au?Subject=Auscope-portal" target="href:blank"><em class="ti-email menu-icon"></em><span>&nbsp;Contact us</span></a></h5>
+    <!-- Portal release version and link to disclaimer -->
+    <h5 mat-menu-item><li appPortalDetailsPanel></li></h5>
 </mat-menu>
